Pre-Final version. ToDo - tests, perfom schemas through OpenAPI docs

+++ fastapi/api/endpoints/login.py
@@ -1,14 +1,15 @@
-from fastapi import APIRouter, Response, HTTPException
+from fastapi import APIRouter, Response, HTTPException, Request
 from sqlalchemy.orm import Session
 from sqlalchemy import and_
 from werkzeug.security import check_password_hash, generate_password_hash
+from random import randint
 
 from ..models import engine, User
-from ..schemas import LoginSchema, CurrentUserResponseModel, ErrorResponseModel, HTTPValidationError
+from ..schemas import LoginSchema, CurrentUserResponseModel, ErrorResponseModel, HTTPValidationError, CodelessErrorResponseModel
 
 login_router = APIRouter()
 
-@login_router.post("/registration")
+@login_router.post("/registration", tags=["auth"])
 def register(data: LoginSchema):
     with Session(bind=engine) as db:
         new_user = User(login=data.login, hash_password=generate_password_hash(data.password))
@@ -27,6 +28,7 @@ def login(response: Response, login_data: LoginSchema):
     if user:
         if check_password_hash(user.hash_password, login_data.password):
             response.set_cookie(key='activeSession', value=login_data.login)
+            response.set_cookie(key='isAdmin', value=user.is_admin)
             return CurrentUserResponseModel(
                 first_name=user.first_name,
                 last_name=user.last_name,
@@ -42,7 +44,11 @@ def login(response: Response, login_data: LoginSchema):
 
 @login_router.get('/logout', responses={
     200: {"description": "Successful Response"}},
-    tags=["auth"], summary="Выход из системы", description="При успешном выходе необходимо удалить установленные Cookies", operation_id="logout_logout_get")
-def logout(response: Response):
+    tags=["auth"], summary="Выход из системы", description="При успешном выходе необходимо удалить установленные Cookies", 
+    operation_id="logout_logout_get")
+def logout(response: Response, request: Request):
+    if not request.cookies.get('activeSession') or not request.cookies.get("isAdmin"):
+        return {"message": "You're already logged out"}
     response.delete_cookie(key='activeSession')
+    response.delete_cookie(key='isAdmin')
     return {"description": "Successful Response"}
